ai-portfolio-system/backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9.6
pymongo>=4.3.3
kafka-python>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
tensorflow>=2.12.0
torch>=2.0.0
pytest>=7.3.1
httpx>=0.24.0
yfinance>=0.2.18
python-dotenv>=1.0.0
pytest-asyncio>=0.21.0
google-cloud-bigquery>=3.10.0
pyspark>=3.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiohttp>=3.8.4 